requirements.txt
moviepy==1.0.3
openai-whisper==20231117
torch==2.2.0
numpy==1.24.3
python-dotenv==1.0.0
ollama==0.1.6 